+++ core/mpc/keysign/referral/build.ts
@@ -1,61 +1,52 @@
 import { create } from '@bufbuild/protobuf'
-import { isChainOfKind } from '@core/chain/ChainKind'
+import { toChainAmount } from '@core/chain/amount/toChainAmount'
 import { AccountCoin } from '@core/chain/coin/AccountCoin'
 import { getCoinBalance } from '@core/chain/coin/balance'
 import { getChainSpecific } from '@core/mpc/keysign/chainSpecific'
-import { FeeSettings } from '@core/mpc/keysign/chainSpecific/FeeSettings'
 import { refineKeysignAmount } from '@core/mpc/keysign/refine/amount'
-import { refineKeysignUtxo } from '@core/mpc/keysign/refine/utxo'
-import { getKeysignUtxoInfo } from '@core/mpc/keysign/utxo/getKeysignUtxoInfo'
 import { MpcLib } from '@core/mpc/mpcLib'
 import { toCommCoin } from '@core/mpc/types/utils/commCoin'
 import { KeysignPayloadSchema } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
 import { WalletCore } from '@trustwallet/wallet-core'
 import { PublicKey } from '@trustwallet/wallet-core/dist/src/wallet-core'
 
-export type BuildSendKeysignPayloadInput = {
+export type BuildReferralKeysignPayloadInput = {
   coin: AccountCoin
-  receiver: string
-  amount: bigint
-  memo?: string
+  memo: string
+  amount: number
   vaultId: string
   localPartyId: string
   publicKey: PublicKey
   libType: MpcLib
   walletCore: WalletCore
-  feeSettings?: FeeSettings
 }
 
-export const buildSendKeysignPayload = async ({
+export const buildReferralKeysignPayload = async ({
   coin,
-  receiver,
-  amount,
   memo,
+  amount,
   vaultId,
   localPartyId,
   publicKey,
-  walletCore,
   libType,
-  feeSettings,
-}: BuildSendKeysignPayloadInput) => {
+  walletCore,
+}: BuildReferralKeysignPayloadInput) => {
   let keysignPayload = create(KeysignPayloadSchema, {
     coin: toCommCoin({
       ...coin,
       hexPublicKey: Buffer.from(publicKey.data()).toString('hex'),
     }),
-    toAddress: receiver,
-    toAmount: amount.toString(),
     memo,
+    toAmount: toChainAmount(amount, coin.decimals).toString(),
     vaultLocalPartyId: localPartyId,
     vaultPublicKeyEcdsa: vaultId,
     libType,
-    utxoInfo: await getKeysignUtxoInfo(coin),
   })
 
   keysignPayload.blockchainSpecific = await getChainSpecific({
     keysignPayload,
-    feeSettings,
     walletCore,
+    isDeposit: true,
   })
 
   const balance = await getCoinBalance(coin)
@@ -66,14 +57,6 @@ export const buildSendKeysignPayload = async ({
     publicKey,
     balance,
   })
-
-  if (isChainOfKind(coin.chain, 'utxo')) {
-    keysignPayload = refineKeysignUtxo({
-      keysignPayload,
-      walletCore,
-      publicKey,
-    })
-  }
 
   return keysignPayload
 }
