+++ core/mpc/keysign/send/build.ts
@@ -1,52 +1,61 @@
 import { create } from '@bufbuild/protobuf'
-import { toChainAmount } from '@core/chain/amount/toChainAmount'
+import { isChainOfKind } from '@core/chain/ChainKind'
 import { AccountCoin } from '@core/chain/coin/AccountCoin'
 import { getCoinBalance } from '@core/chain/coin/balance'
 import { getChainSpecific } from '@core/mpc/keysign/chainSpecific'
+import { FeeSettings } from '@core/mpc/keysign/chainSpecific/FeeSettings'
 import { refineKeysignAmount } from '@core/mpc/keysign/refine/amount'
+import { refineKeysignUtxo } from '@core/mpc/keysign/refine/utxo'
+import { getKeysignUtxoInfo } from '@core/mpc/keysign/utxo/getKeysignUtxoInfo'
 import { MpcLib } from '@core/mpc/mpcLib'
 import { toCommCoin } from '@core/mpc/types/utils/commCoin'
 import { KeysignPayloadSchema } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
 import { WalletCore } from '@trustwallet/wallet-core'
 import { PublicKey } from '@trustwallet/wallet-core/dist/src/wallet-core'
 
-export type BuildReferralKeysignPayloadInput = {
+export type BuildSendKeysignPayloadInput = {
   coin: AccountCoin
-  memo: string
-  amount: number
+  receiver: string
+  amount: bigint
+  memo?: string
   vaultId: string
   localPartyId: string
   publicKey: PublicKey
   libType: MpcLib
   walletCore: WalletCore
+  feeSettings?: FeeSettings
 }
 
-export const buildReferralKeysignPayload = async ({
+export const buildSendKeysignPayload = async ({
   coin,
-  memo,
+  receiver,
   amount,
+  memo,
   vaultId,
   localPartyId,
   publicKey,
-  libType,
   walletCore,
-}: BuildReferralKeysignPayloadInput) => {
+  libType,
+  feeSettings,
+}: BuildSendKeysignPayloadInput) => {
   let keysignPayload = create(KeysignPayloadSchema, {
     coin: toCommCoin({
       ...coin,
       hexPublicKey: Buffer.from(publicKey.data()).toString('hex'),
     }),
+    toAddress: receiver,
+    toAmount: amount.toString(),
     memo,
-    toAmount: toChainAmount(amount, coin.decimals).toString(),
     vaultLocalPartyId: localPartyId,
     vaultPublicKeyEcdsa: vaultId,
     libType,
+    utxoInfo: await getKeysignUtxoInfo(coin),
   })
 
   keysignPayload.blockchainSpecific = await getChainSpecific({
     keysignPayload,
+    feeSettings,
     walletCore,
-    isDeposit: true,
   })
 
   const balance = await getCoinBalance(coin)
@@ -57,6 +66,14 @@ export const buildReferralKeysignPayload = async ({
     publicKey,
     balance,
   })
+
+  if (isChainOfKind(coin.chain, 'utxo')) {
+    keysignPayload = refineKeysignUtxo({
+      keysignPayload,
+      walletCore,
+      publicKey,
+    })
+  }
 
   return keysignPayload
 }
